restore dev checkbox
